// src/models/peritacion.ts
export class Peritacion {

    id: number;
    aseguradora: number;
    inspeccion: number;
    dominio: string;
    nro_siniestro: string;
// ...
    provincia: number;
    localidad: number;
    calle: string;
    numero: string;
    depto: string;
    entre_calles: string;
    prod_nombre: string;
    prod_apellido: string;
    prod_email: string;
    prod_telefono: string;
    prod_cuit: string;
    estado: number;
// ...
}